+++ CHANGELOG.md
@@ -8,9 +8,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 #### Added
 
+- Added Vizier and Gaia source search along with corresponding tests [#607](https://github.com/askap-vast/vast-tools/pull/607)
 - Allow recalculation of sources_df without pairs existing [#605](https://github.com/askap-vast/vast-tools/pull/605)
 
 #### Changed
+
+- Cleaned up angular offset handling for SIMBAD and NED source search [#607](https://github.com/askap-vast/vast-tools/pull/607)
 
 #### Fixed
 
@@ -22,6 +25,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 #### List of PRs
 
+- [#607](https://github.com/askap-vast/vast-tools/pull/607): feat, fix: Improve external source search queries
 - [#611](https://github.com/askap-vast/vast-tools/pull/611): fix: Fixed SkyView connection hanging
 - [#609](https://github.com/askap-vast/vast-tools/pull/609: fix: Fixed inaccurate planet crossmatching by changing solar system ephemeris to JPL de432s
 - [#605](https://github.com/askap-vast/vast-tools/pull/605): feat: Allow recalculation of sources_df without pairs existing